requirements.txt
Pillow==9.4.0
requests==2.27.1
streamlit==1.10.0
streamlit_option_menu==0.3.2
